+++ src/app/lib/lforms-widget-registry.ts
@@ -75,6 +75,7 @@ import {
 } from "./widgets/expression-editor/calculated-initial-expression/calculated-initial-expression.component";
 import {EnableWhenExpressionComponent} from "./widgets/expression-editor/enable-when-expression/enable-when-expression.component";
 import {InitialComponent} from "./widgets/initial/initial.component";
+import {MinMaxOccursComponent} from './widgets/min-max-occurs/min-max-occurs.component';
 import {ExtensionUrlComponent} from './widgets/extension-url/extension-url.component';
 
 @Injectable()
@@ -147,6 +148,8 @@ export class LformsWidgetRegistry extends DefaultWidgetRegistry {
     this.register('coding-display', CodingDisplayComponent);
     this.register('enable-when-method', EnableWhenMethodComponent);
     this.register('enable-when-expression', EnableWhenExpressionComponent);
+    this.register('min-max-occurs', MinMaxOccursComponent);
+
     this.register('meta-profile', MetaProfileComponent);
     this.register('usage-context', UsageContextComponent);
     this.register('usage-context-code', UsageContextCodeComponent);