+++ CHANGELOG.md
@@ -2,12 +2,21 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [12.4.0] 2026-08-26
+## [12.5.0] 2026-09-14
 ### Added
 - Added support for the `attachment` item data type. This includes the "Maximum size"
   restriction (entered with a Bytes/KB/MB/GB helper and stored in bytes) and the
   "Mime type" restriction, as well as an initial Attachment value (URL, file upload and
   base64Binary).
+
+## [12.4.1] 2026-09-10
+### Added
+- Added `SECURITY.md`
+
+## [12.4.0] 2026-08-26
+### Added
+- Added support for `questionnaire-minOccurs` and `questionnaire-maxOccurs` extensions
+  via a combined Min/Max Occurs widget that is shown when an item allows repeats.
 
 ## [12.3.1] 2026-08-20
 ### Fixed